Refactored creating tree
