+++ src/components/fileNode.tsx
@@ -4,127 +4,12 @@ import "../styles.scss";
 import { IoIosArrowDown, IoIosArrowUp } from "react-icons/io";
 import { AiFillFolder, AiFillFolderOpen } from "react-icons/ai";
 
-import getAliases from "../parser/getAliases";
-import ErrorMessage from "./errorMessage";
 
 
-// filenode component takes in children and filename
-const FileNode = (props) => {
+const fileNode = (props) => {
 
-  let childList = [];
+  console.log('children:', props.children)
   const [showChildren, setShow] = useState(false);
-  let components = props.totalComponents;
-  
-
-  let aliases = props.aliases;
-
-  // creates the tree; file is props
-  function createTree() {
-
-    for (let i = 0; i < props.children.length; i++) {
-
-      console.log('currently on:', props.children[i])
-
-      // loop to generate aliases from out script tag
-      if (props.children[i].type !== "svelteComponent" && props.children[i].type !== "svelteElement" && props.children[i].type !== 'svelteScript') {
-        console.log('Filtered out file: ', props.children[i]);
-        continue;
-      }
-
-      
-      if (props.children[i].type === "svelteScript") {
-        aliases = getAliases(props.children[i]);
-        if (aliases === undefined){
-          aliases = props.aliases;
-        }
-        
-          
-        
-        console.log(`Updated aliase to: `, aliases);
-        continue;
-      }
-
-      //recursion for elements like main,p,h1
-      else if (props.children[i].type === "svelteElement") {
- 
-        childList.push(
-          <FileNode
-            children={props.children[i].children}
-            fileName={props.children[i].tagName}
-            fileType={"svelteElement"}
-            svelteFiles={props.svelteFiles}
-            aliases = {aliases}
-            setTotalComponents={props.setTotalComponents} 
-            totalComponents={props.totalComponents}
-            errorLog = {props.errorLog}
-            setError = {props.setError}
-          />
-        );
-        console.log('Pushed element to child list: ', childList);
-
-      } 
-      else if(props.children[i].type === "svelteComponent"){
-        
-        let searchStr;
-        let hasAlias = false;
-        if(aliases[props.children[i].tagName]){
-          hasAlias = true;
-          searchStr = aliases[props.children[i].tagName];
-        } 
-
-        // if it cannot find an alias for the component throw an error
-        if(!hasAlias){
-          props.setError([...props.errorLog, <ErrorMessage errorCode={404} errorMessage={`Failed to Find Import for ${props.children[i].tagName}`}/>])
-          continue;
-        }
-        
-      
-          for (let j = 0; j < props.svelteFiles.length; j++) {
-            console.log('searching svelte files....', props.children[i]);
-
-            const string = props.svelteFiles[j].fileName;
-
-
-            if ( string === searchStr) {
-              console.log('Found!')
-
-              components += 1;
-               childList.push(
-                
-                <FileNode 
-                children={props.svelteFiles[j].children} 
-                fileName={props.children[i].tagName} 
-                fileType={"svelteComponent"} 
-                svelteFiles={props.svelteFiles} 
-                aliases={aliases}
-                setTotalComponents={props.setTotalComponents} 
-                totalComponents={props.totalComponents}
-                errorLog = {props.errorLog}
-                setError = {props.setError}
-                />
-                
-              );
-
-              console.log('Pushed with new alias');
-              break;
-             
-            }
-          }
-        }
-      }
-    }
-    //props.setTotalComponents(props.totalComponents + components);
-
-  
-
-
-  // recursion to continuously find children
-  // if there are children, continuously invoke createTree on line 21
-  if (props.children) {
-    createTree()
-  }
-
-  // if there are no children left to parse, render/return. Don't need to create/continue tree if no children
 
   return (
     <div id="treeNode">
@@ -133,8 +18,8 @@ const FileNode = (props) => {
         props.fileType === "svelteElement" ? (
           <div>
             {
-              // if there is a length, if there is children
-              props.children.length ? (
+              // if there are children
+              props.children ? (
                 // if current state of showchildren is truthy
                 showChildren ? (
                   // create up arrow icon
@@ -165,17 +50,17 @@ const FileNode = (props) => {
 
             <div id="childrenNodes">
               {/* if showchildren is truthy, show childlist, otherwise show nothing */}
-              {showChildren ? childList : null}
+              {showChildren ? props.children : null}
             </div>
           </div>
         ) : (
-          //otherwise do this stuff
+          // component rendering
           <div>
             <div id="componentNode">
               Component
               <div>Name: {props.fileName}</div>
               <div>Children: {props.children ? props.children.length : 0}</div>
-              {props.children.length ? (
+              {props.children ? (
                 showChildren ? (
                   <button
                     id="expandButton"
@@ -194,11 +79,12 @@ const FileNode = (props) => {
               ) : null}
             </div>
 
-            <div id="childrenNodes">{showChildren ? childList : null}</div>
+            <div id="childrenNodes">{showChildren ? props.children : null}</div>
           </div>
         )
       }
     </div>
   );
-};
-export default FileNode;
+}
+
+export default fileNode;
